auction_bid_tracker/requirements.txt
Django==3.0.1
djangorestframework==3.11.0
gunicorn
